+++ docs/changelog.md
@@ -1,6 +1,9 @@
 # TBD
 ### Features
 * Port over the relevant bits of documentation from `kurtosis-libs`
+* Added functionality to the API container's API and `NetworkContext` for:
+    * Registering static files with the API container
+    * Using those static files when starting a service
 
 # 0.2.2
 ### Changes